backend/requirements.txt
# Core FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain & AI - Compatible versions
langchain==0.2.11
langchain-openai==0.1.19
langchain-community==0.2.10
langchain-chroma==0.1.2

# Document Processing
pypdf==3.17.4
chromadb==0.4.18

# Environment & Config
python-dotenv==1.0.0
pydantic==2.8.2

# Search & Retrieval - Updated for compatibility
scikit-learn==1.3.2

# OpenAI
openai==1.35.13

# AWS SDK (for production)
boto3==1.34.0

# Rate Limiting & Caching (NEW)
redis==5.0.1
hiredis==2.2.3

# Monitoring (optional)
prometheus-client==0.19.0
psutil==5.9.6

# HTTP requests
httpx==0.25.2

# JWT Authentication
jwt==1.3.1

sentence-transformers>=2.2.0